+++ src/Beef.Data.Cosmos/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 Represents the **NuGet** versions.
 
+## v3.1.x
+- *Enhancement:* `CosmosDbMapper.CreateArgs` method extended to support an optional `action` to further update the `CosmosDbArgs` (helps simplify the generated code logic).
+
 ## v3.1.4
 - *Enhancement:* All references to `DateTime.Now` have been updated to `Cleaner.Clean(DateTime.Now)`.